Remove dead code and rename misnamed API client
- Delete mock-chat.ts (86 lines, zero imports — entirely dead code)
- Rename mock-artifacts.ts → artifacts-api.ts (it's the real API client, not a mock)
- Update all import references (App.tsx, Desktop.tsx, ArtifactIcon.tsx)
- Update design doc reference

Part of #6
